+++ src/index.ts
@@ -2,6 +2,14 @@ export {Task} from './model/task';
 export {TaskState} from './model/taskState';
 export {Topic} from './model/topic';
 
+export {HeijunkaBoard} from './heijunka/HeijunkaBoard';
+export {KanbanCard} from './heijunka/KanbanCard';
+export {Project} from './heijunka/Project';
+export {State} from './heijunka/State';
+export {StateHistory} from './heijunka/StateHistory';
+export {StateModel} from './heijunka/StateModel';
+export {StateTransition,TransitionType} from './heijunka/StateTransition';
+
 export {ObjectEvent} from './objectEvents/objectEvent'
 export {ObjectEventFactory} from './objectEvents/objectEventFactory'
 export {ProcessObjectEventCommand} from './objectEvents/processObjectEventCommand'